backend/pyproject.toml
[project]
name = "modnet-finetune-backend"
version = "0.1.0"
description = "MODNet fine-tuning backend with FastAPI"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "onnx>=1.16.0",
    "onnxruntime>=1.19.0",
    "onnxconverter-common>=1.14.0",
    "datasets>=3.0.0",
    "huggingface-hub>=0.26.0",
    "Pillow>=10.0.0",
    "scipy>=1.14.0",
    "scikit-image>=0.24.0",
    "numpy>=1.26.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "aiofiles>=24.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
]
